+++ hs/src/HelVM/HelPA/Assembler/Macro.hs
@@ -4,9 +4,6 @@ import           HelVM.HelPA.Assembler.Value
 
 data Macro a =
     Micro a
-  | Call Header
-  | Def Header [Macro a]
-  deriving stock (Eq , Read , Show)
-
-data Header = Header [IntegerValue] Text
+  | Call [IntegerValue] Identifier
+  | Def [Identifier] Identifier [Macro a]
   deriving stock (Eq , Read , Show)